--- src/main/java/coupon/DataSourceType.java
@@ -1,5 +1,0 @@
-package coupon;
-
-public enum DataSourceType {
-    READER,WRITER
-}
