+++ src/main/java/coupon/data/config/DataSourceType.java
@@ -1,0 +1,5 @@
+package coupon.data.config;
+
+public enum DataSourceType {
+    READER, WRITER
+}
